Update skill-desc.csv (Marune)
Some things to note:
'わずかに', '少し' and skills without neither of these words have different levels of effectiveness. I divided each in the following way:
速度がわずかに上がる　→　Slightly increase speed
速度が少し上がる　→　Lightly increase speed / Increase speed a little / Increase speed a bit
速度が上がる　→　Increase speed
Not dividing this way would make certain skills seem more stronger than others, when in reality, it only increases 'slightly' (わずかに) or 'more than slightly' (少し).
It's easier in Japanese to find these keywords, but in English they seem to be in different spots, and not always on the last part of the sentence.